adding debug and verbose options via argparse

+++ alcor.py
@@ -1,34 +1,47 @@
 #!/usr/local/python/bin/python
-# Script to observe with the all-sky camera
-# 
-# to do:	
-#	add basic camera control	
+#
+# Description:
+# Script to observe with the Alcor all-sky camera
+# This script applies one exptime during the day
+# and another for night time. 
+#
+# The main day/night settings are set in the 
+# config files in the Git repo. Check there for 
+# device and image settings.
+#
+# To do:	
 #	add Pyro support
 #	add archiving
-#	work out when to jump from day to night mode using el_sun
-#	run the camera in one thread 
-#	run image analysis in another
-# 	
-# The main settings are set in the config file
-# given below. Check here for device and image
-# settings
+#	run image analysis in another subprocess?
 #
+
 import os, time
 import signal, subprocess
 from astropy.time import Time
 from astropy.coordinates import EarthLocation, AltAz
 from astropy.coordinates import get_sun
 from datetime import datetime as dt
+import argparse as ap
 
+# edit here
 live_image="/home/ops/webcam/allsky.jpeg"
 sun_alt_limit = -5
-die=False
 
 # observatory set up
 olat=-24.-(37./60.)-(38./3600.)
 olon=-70.-(24./60.)-(15./3600.)
 elev=2418.
 paranal=EarthLocation(lat=olat*u.deg,lon=olon*u.deg,height=elev*u.m)
+die=False
+
+# parse command line
+def argParse():
+	parser=ap.ArgumentParser(description='Alcor all-sky camera control script. Edit camera\'s day/night settings using the .conf files in the Git repo.')
+	parser.add_argument("--debug",help="run in deugging mode",action="store_true")
+	parser.add_argument("--v",help="increased verbosity",action="store_true")
+	return parser.parse_args()
+
+args=argParse()
 
 # work out if it is day or night time
 # based on the current Sun altitude
@@ -38,39 +51,51 @@ def dayOrNight():
 	sunaltaz = get_sun(time).transform_to(altazframe)
 	sunalt=sunaltaz.alt.deg[0]
 	if sunalt <= sun_alt_limit:
-		return 'night'
+		result='night'
 	else:
-		return 'day'
+		result='day'
+	if args.v:
+		print "Sun altitude: %.2f - %s" % (sunalt,result)
+	return result
 
 # set up Ctrl+C handling
-def signal_handler(signal,frame):
+def signalHandler(signal,frame):
 	global die
 	print "Ctrl+C caught, exiting..."
 	die=True
-signal.signal(signal.SIGINT,signal_handler)
+signal.signal(signal.SIGINT,signalHandler)
 
 # main function
 def main():
 	global die
 	don=dayOrNight()
 	comm='fswebcam -c paranal_%s.conf' % (don)
-	print "%s - %s" % (dt.utcnow(),comm)
-	pro=subprocess.Popen(comm,subprocess.PIPE,shell=True,preexec_fn=os.setsid)
+	if args.v:
+		print "%s - %s" % (dt.utcnow(),comm)
+	if not args.debug:
+		pro=subprocess.Popen(comm,subprocess.PIPE,shell=True,preexec_fn=os.setsid)
 	while(1):
 		don_new=dayOrNight()
+		# change from day to night, and vice versa
 		if don_new != don:
-			# change settings
-			print "%s - Changing the settings to %s" % (dt.utcnow(),don)
-			os.killpg(os.getpgid(pro.pid), signal.SIGTERM)
+			if args.v:
+				print "%s - Changing the settings to %s" % (dt.utcnow(),don)
+			if not args.debug:
+				os.killpg(os.getpgid(pro.pid), signal.SIGTERM)
 			time.sleep(30)
 			comm='fswebcam -c paranal_%s.conf' % (don_new)
-			print "%s - %s" % (dt.utcnow(),comm)
-			pro=subprocess.Popen(comm,subprocess.PIPE,shell=True,preexec_fn=os.setsid)
+			if args.v:
+				print "%s - %s" % (dt.utcnow(),comm)
+			if not args.debug:
+				pro=subprocess.Popen(comm,subprocess.PIPE,shell=True,preexec_fn=os.setsid)
 			don=don_new
+		# wait for a bit between checks for change in day/night
 		time.sleep(10)
+		# if ctrl+c, die correctly
 		if die == True:
 			print "Killing fswebcam"
-			os.killpg(os.getpgid(pro.pid), signal.SIGTERM)
+			if not args.debug:
+				os.killpg(os.getpgid(pro.pid), signal.SIGTERM)
 			print "Exiting..."
 			break
